+++ apps/agent/pkg/membership/redis.go
@@ -9,8 +9,10 @@ import (
 	"time"
 
 	"github.com/redis/go-redis/v9"
+	clusterv1 "github.com/unkeyed/unkey/apps/agent/gen/proto/cluster/v1"
 	"github.com/unkeyed/unkey/apps/agent/pkg/events"
 	"github.com/unkeyed/unkey/apps/agent/pkg/logging"
+	"github.com/unkeyed/unkey/apps/agent/pkg/repeat"
 )
 
 type membership struct {
@@ -25,6 +27,9 @@ type membership struct {
 	heartbeatLock   sync.Mutex
 	membersSnapshot []Member
 
+	syncTtl       time.Duration
+	syncFrequency time.Duration
+
 	nodeId string
 }
 
@@ -33,6 +38,16 @@ type Config struct {
 	NodeId   string
 	RpcAddr  string
 	Logger   logging.Logger
+
+	// How frequently to heartbeat to redis to indicate that this node is still alive as well as to
+	// download the list of other nodes.
+	SyncFrequency time.Duration
+
+	// The time to live for the heartbeat key in redis. If this node fails to heartbeat within this
+	// time, it will be considered dead.
+	//
+	// Set this to roughly 2x the sync frequency or at least 5s longer than the sync frequency.
+	SyncTtl time.Duration
 }
 
 func New(config Config) (Membership, error) {
@@ -41,6 +56,8 @@ func New(config Config) (Membership, error) {
 	if err != nil {
 		return nil, fmt.Errorf("failed to parse redis url: %w", err)
 	}
+	opts.MaxRetryBackoff = time.Second
+	opts.MaxRetries = 5
 
 	rdb := redis.NewClient(opts)
 
@@ -49,6 +66,12 @@ func New(config Config) (Membership, error) {
 		return nil, fmt.Errorf("failed to ping redis: %w", err)
 	}
 
+	if config.SyncTtl == 0 {
+		config.SyncTtl = 15 * time.Second
+	}
+	if config.SyncFrequency == 0 {
+		config.SyncFrequency = 10 * time.Second
+	}
 	return &membership{
 		started:         false,
 		logger:          config.Logger,
@@ -59,8 +82,14 @@ func New(config Config) (Membership, error) {
 		nodeId:          config.NodeId,
 		heartbeatLock:   sync.Mutex{},
 		membersSnapshot: []Member{},
-	}, nil
 
+		syncTtl:       config.SyncTtl,
+		syncFrequency: config.SyncFrequency,
+	}, nil
+}
+
+func (m *membership) NodeId() string {
+	return m.nodeId
 }
 
 func (m *membership) Join(addrs ...string) (int, error) {
@@ -73,15 +102,17 @@ func (m *membership) Join(addrs ...string) (int, error) {
 	m.started = true
 	m.logger.Info().Msg("Initilizing redis membership")
 
-	m.heartbeat()
+	err := m.Sync()
+	if err != nil {
+		return 0, fmt.Errorf("failed to sync: %w", err)
+	}
 
-	go func() {
-		t := time.NewTicker(10 * time.Second)
-		defer t.Stop()
-		for range t.C {
-			m.heartbeat()
+	repeat.Every(m.syncFrequency, func() {
+		err := m.Sync()
+		if err != nil {
+			m.logger.Error().Err(err).Msg("failed to sync")
 		}
-	}()
+	})
 
 	members, err := m.Members()
 	if err != nil {
@@ -95,33 +126,32 @@ func (m *membership) heartbeatRedisKey() string {
 	return fmt.Sprintf("cluster::membership::nodes::%s", m.nodeId)
 }
 
-func (m *membership) heartbeat() {
+func (m *membership) Sync() error {
 	m.heartbeatLock.Lock()
 	defer m.heartbeatLock.Unlock()
 	b, err := json.Marshal(Member{
 		Id:      m.nodeId,
 		RpcAddr: m.rpcAddr,
+		State:   clusterv1.NodeState_NODE_STATE_ACTIVE.String(),
 	})
 	if err != nil {
-		m.logger.Err(err).Msg("failed to marshal self")
-		return
+		return fmt.Errorf("failed to marshal self: %w", err)
 	}
-	err = m.rdb.Set(context.Background(), m.heartbeatRedisKey(), string(b), 15*time.Second).Err()
+	err = m.rdb.Set(context.Background(), m.heartbeatRedisKey(), string(b), m.syncFrequency).Err()
 	if err != nil {
 		m.logger.Error().Err(err).Msg("failed to set node")
 	}
 
 	newMembers, err := m.Members()
 	if err != nil {
-		m.logger.Error().Err(err).Msg("failed to get members")
-		return
+		return fmt.Errorf("failed to get members: %w", err)
 	}
 
 	// Check for left nodes
 	for _, oldMember := range m.membersSnapshot {
 		found := false
 		for _, newMember := range newMembers {
-			if oldMember.Id == newMember.Id {
+			if oldMember.Id == newMember.Id && newMember.State == clusterv1.NodeState_NODE_STATE_ACTIVE.String() {
 				found = true
 				break
 			}
@@ -140,17 +170,32 @@ func (m *membership) heartbeat() {
 				break
 			}
 		}
-		if !found {
+		if !found && newMember.State == clusterv1.NodeState_NODE_STATE_ACTIVE.String() {
 			m.joinEvents.Emit(newMember)
 		}
 
 	}
 	m.membersSnapshot = newMembers
+	return nil
 
 }
 
 func (m *membership) Leave() error {
-	return m.rdb.Del(context.Background(), m.heartbeatRedisKey()).Err()
+
+	b, err := json.Marshal(Member{
+		Id:      m.nodeId,
+		RpcAddr: m.rpcAddr,
+		State:   clusterv1.NodeState_NODE_STATE_LEAVING.String(),
+	})
+	if err != nil {
+		return fmt.Errorf("failed to marshal self: %w", err)
+	}
+
+	err = m.rdb.Set(context.Background(), m.heartbeatRedisKey(), string(b), m.syncFrequency).Err()
+	if err != nil {
+		return fmt.Errorf("failed to set node: %w", err)
+	}
+	return nil
 }
 
 func (m *membership) Members() ([]Member, error) {