+++ apps/agent/services/vault/keyring/roll_keys.go
@@ -3,19 +3,28 @@ package keyring
 import (
 	"context"
 	"fmt"
+
+	"github.com/unkeyed/unkey/apps/agent/pkg/tracing"
+	"github.com/unkeyed/unkey/apps/agent/services/vault/storage"
 )
 
 func (k *Keyring) RollKeys(ctx context.Context, ringID string) error {
+	ctx, span := tracing.Start(ctx, tracing.NewSpanName("keyring", "RollKeys"))
+	defer span.End()
 	lookupKeys, err := k.store.ListObjectKeys(ctx, k.buildLookupKey(ringID, "dek_"))
 	if err != nil {
 		return fmt.Errorf("failed to list keys: %w", err)
 	}
 
 	for _, objectKey := range lookupKeys {
-		b, err := k.store.GetObject(ctx, objectKey)
+		b, found, err := k.store.GetObject(ctx, objectKey)
 		if err != nil {
 			return fmt.Errorf("failed to get object: %w", err)
 		}
+		if !found {
+			return storage.ErrObjectNotFound
+		}
+
 		dek, encryptionKeyId, err := k.DecodeAndDecryptKey(ctx, b)
 		if err != nil {
 			return fmt.Errorf("failed to decode and decrypt key: %w", err)