+++ apps/api/src/pkg/keys/service.ts
@@ -48,7 +48,13 @@ type NotFoundResponse = {
 type InvalidResponse = {
   valid: false;
   publicMessage?: string;
-  code: "FORBIDDEN" | "RATE_LIMITED" | "USAGE_EXCEEDED" | "DISABLED" | "INSUFFICIENT_PERMISSIONS";
+  code:
+    | "FORBIDDEN"
+    | "RATE_LIMITED"
+    | "EXPIRED"
+    | "USAGE_EXCEEDED"
+    | "DISABLED"
+    | "INSUFFICIENT_PERMISSIONS";
   key: Key;
   api: Api;
   ratelimit?: {
@@ -359,7 +365,13 @@ export class KeyService {
     const expires = data.key.expires ? new Date(data.key.expires).getTime() : undefined;
     if (expires) {
       if (expires < Date.now()) {
-        return Ok({ valid: false, code: "NOT_FOUND" });
+        return Ok({
+          valid: false,
+          code: "EXPIRED",
+          key: data.key,
+          api: data.api,
+          permissions: data.permissions,
+        });
       }
     }
 
@@ -540,6 +552,7 @@ export class KeyService {
     const res = await this.rateLimiter.multiLimit(
       c,
       Object.values(ratelimits).map((r) => ({
+        name: r.name,
         async: !!key.ratelimitAsync,
         workspaceId: key.workspaceId,
         identifier: key.id,
@@ -556,6 +569,13 @@ export class KeyService {
       });
 
       return [false, undefined];
+    }
+
+    if (res.val.triggered !== null) {
+      /**
+       * This is undocumented and only used internally for test assertions
+       */
+      c.res.headers.set("Unkey-Ratelimit-Triggered", res.val.triggered);
     }
 
     return [