+++ apps/dashboard/app/(app)/settings/root-keys/[keyId]/layout.tsx
@@ -35,12 +35,12 @@ export default async function Layout({ children, params: { keyId } }: Props) {
   return (
     <div className="flex flex-col gap-4">
       <Card>
-        <CardContent className="flex justify-between flex-wrap divide-x">
+        <CardContent className="flex flex-wrap justify-between divide-x">
           <Metric label="ID" value={<span className="font-mono">{key.id}</span>} />
           <Metric label="Name" value={key.name ?? "-"} />
           <Metric label="Created At" value={key.createdAt.toDateString()} />
           <Metric
-            label={key.expires && key.expires.getTime() < Date.now() ? "Expired" : "Expires"}
+            label={key.expires && key.expires.getTime() < Date.now() ? "Expired" : "Expires in"}
             value={key.expires ? ms(key.expires.getTime() - Date.now()) : "-"}
           />