+++ packages/cache/src/stores/index.ts
@@ -1,3 +1,4 @@
 export * from "./cloudflare";
 export * from "./interface";
 export * from "./memory";
+export * from "./upstash-redis";