+++ apps/planetfall/CHANGELOG.md
@@ -1,5 +1,25 @@
 # planetfall
 
+## 0.1.13
+
+### Patch Changes
+
+- @unkey/api@0.22.1
+
+## 0.1.12
+
+### Patch Changes
+
+- Updated dependencies [53a1df1]
+  - @unkey/api@0.22.0
+
+## 0.1.11
+
+### Patch Changes
+
+- Updated dependencies [9dab761]
+  - @unkey/api@0.21.0
+
 ## 0.1.10
 
 ### Patch Changes